modularize and improve booklist and modal UI
