--- src/components/BookSearch/common/common-styles/borrow-button.css
@@ -1,3 +1,0 @@
-.border-radius {
-    border-radius: 150px !important;
-}
