--- src/components/styles/book-detail.css
@@ -1,11 +1,0 @@
-.upper-radius {
-  border-radius: 25px 25px 0px 0px;
-}
-
-.bottom-radius {
-  border-radius: 0px 0px 25px 25px;
-}
-
-.gray-text {
-  color: rgb(167, 167, 167);
-}
